Force rebuild when requested. Use Makefile to pack upx.

+++ docker/windows-debug/_build.sh
@@ -5,6 +5,10 @@ BRANCH=master
 if [ "$1xxx" != "xxx" ] ; then
 	BRANCH=$1
 fi
+REBUILD=0
+if [ "$2xxx" == "rebuildxxx" ] ; then
+	REBUILD=1
+fi
 NAME=nitrokey-app.$BRANCH
 export PATH=$PATH:/mxe/usr/bin
 
@@ -17,27 +21,21 @@ git pull
 git checkout $BRANCH
 git submodule update --init --recursive
 
-for i in build-cmake build-qt; do
-	mkdir -p $i
-	rm -rf $i/*
-done
+if [ $REBUILD -eq 1 ]; then
+	for i in build-qt-release build-qt; do
+		mkdir -p $i
+		rm -rf $i/*
+	done
+fi
 
-# pushd build-cmake
-#i686-w64-mingw32.static.posix-cmake ..
-#make -j4
-#cp nitrokey-app.exe /build/nitrokey-app-cmake.exe
-#popd
-#
-#pushd build-cmake
-#export PATH=$PATH:/build/mxe/bin
-#i686-w64-mingw32.static.posix-cmake .. -DLOG_VOLATILE_DATA=ON
-#make -j4
-#cp nitrokey-app.exe /build/nitrokey-app-cmake-volatile.exe
-#popd
+function mkdircd(){
+	mkdir -p $1
+	pushd $1
+}
 
 CONFIG=qt-volatile
 FNDMG=nitrokey-app-`git describe`-$BRANCH.$CONFIG.exe
-pushd build-qt
+mkdircd build-qt
 i686-w64-mingw32.static.posix-qmake-qt5 .. CONFIG+=CONSOLE DEFINES+=LOG_VOLATILE_DATA
 make -j4 -f Makefile.Release
 cp release/nitrokey-app.exe /build/$FNDMG
@@ -45,7 +43,7 @@ popd
 
 CONFIG=qt-release
 FNDMG=nitrokey-app-`git describe`-$BRANCH.$CONFIG.exe
-pushd build-qt
+mkdircd build-qt-release
 i686-w64-mingw32.static.posix-qmake-qt5 ..
 make -j4 -f Makefile.Release
 cp release/nitrokey-app.exe /build/$FNDMG
@@ -54,9 +52,7 @@ popd
 
 cd /build/
 mkdir -p upx
-#rm -rf upx/*
 
 for i in `ls *exe`; do
-	upx $i -o upx/$i -l
-	upx -t upx/$i
+	make -f /app/Makefile.upx upx/$i
 done
